Capture total and sampled span count metrics - refactored and moved the following from metricHelper   * renamed metrics to metrics_helper   * metrics - register and unregister latency histogram,     trace count and span count metrics   * metrics_exporter - start / stop prometheus exporter
