+++ assertsprocessor/trace.go
@@ -6,6 +6,17 @@ type trace struct {
 	segments []*traceSegment
 }
 
+func (tr *trace) getSpanCount() int {
+	var count int
+	for _, ts := range tr.segments {
+		count += len(ts.entrySpans) + len(ts.exitSpans) + len(ts.internalSpans)
+		if ts.rootSpan != nil {
+			count += 1
+		}
+	}
+	return count
+}
+
 type traceSegment struct {
 	resourceSpans    *ptrace.ResourceSpans
 	namespace        string
